+++ components/certifit-checkout/index.tsx
@@ -99,8 +99,7 @@ export default function CertifitCheckoutComponent({ courseCanonicalId }: Certifi
         </>
     );
 }
-// userCourse 2
-// card 10
+
 type ModalPurchaseResultProp = {
     result: {
         success: boolean,